refactor(db-core): db_core_service.py를 client/facade/dbapi-shim 3계층으로 분리
거대해진 db_core_service.py(782줄)를 역할별로 분리했다. 원본 파일은 기존 공개
이름을 전부 from-import하는 순수 재수출 모듈로 축소해 기존 import 경로 13곳을
그대로 유지한다.

- db_core_client.py: JSONL 클라이언트(DbCoreServiceClient) + 엔진/버전 헬퍼
- db_core_facade.py: DbEndpoint + DbCoreFacade + 공유 facade 라이프사이클
  (atexit 등록은 이 모듈에만 존재 — 재수출 모듈 이중 등록 방지)
- db_core_dbapi_shim.py: RustDbConnector/RustDbConnection/RustDbCursor

부수 정리:
- create_rust_db_connector가 DbEndpoint를 1회만 생성해 RustDbConnector에
  keyword-only endpoint로 전달, 이중 repackaging 제거 (레거시 7-param
  위치 인자 시그니처는 그대로 유지)
- RustDbCursor.execute가 execute_on_connection_result를 무조건 호출하도록
  단순화(hasattr 분기 + rows_affected fallback 제거)
- get_tables/select_db의 수동 DbEndpoint 재구성을 dataclasses.replace로 교체
- 분할 후 무효화되는 모듈 객체 monkeypatch 3곳을 db_core_dbapi_shim 대상으로
  이동 + 재수출 회귀 테스트 신설

covers: CC-001, CC-006, CC-007, CC-014

검증: 신규/수정 4개 py_compile 통과, 대상 pytest 3개 파일 91개 통과,
전체 pytest 1811개 중 1810개 통과(macOS validation 1건은 GitHub CI 의존
기존 known-flaky).
